refactor(dto): Flatten remote.dto into data.dto for clarity and clean architecture alignment – Progresses #5
This commit renames the `remote.dto` package to `dto` directly under `data/` to improve readability and better align with clean architecture conventions.

- Simplified package name from `remote.dto` to `dto`
- Updated all relevant imports across:
  - Mapper
  - Repository
  - ViewModel
- Removed unnecessary nesting (`remote`) as DTO context is already scoped within `data` layer

This change improves modular clarity and aligns with clean architecture standards.

Progresses #5
