+++ app/src/main/java/com/sadaquekhan/justeatassessment/ui/components/SearchBar.kt
@@ -1,4 +1,4 @@
-package com.sadaquekhan.justeatassessment.ui.screen.components
+package com.sadaquekhan.justeatassessment.ui.components
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
